ui: make links clickable in public notes Ref: #1838
