+++ ui/src/common/components/PublicNotesList/PublicNotesList.scss
@@ -1,0 +1,3 @@
+.__PublicNotesList__ {
+  overflow-wrap: break-word;
+}
