+++ ui/src/common/components/PublicNotesList/index.js
@@ -1,0 +1,3 @@
+import PublicNotesList from './PublicNotesList';
+
+export default PublicNotesList;
